+++ app/src/main/java/io/mapwize/mapwizeuicomponents/MainActivity.kt
@@ -2,20 +2,19 @@ package io.mapwize.mapwizeuicomponents
 
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import android.util.Log
 import com.mapbox.mapboxsdk.maps.MapboxMap
 import io.indoorlocation.core.IndoorLocation
 import io.indoorlocation.manual.ManualIndoorLocationProvider
 import io.mapwize.mapwizecomponents.ui.MapwizeFragment
 import io.mapwize.mapwizecomponents.ui.MapwizeFragmentUISettings
-import io.mapwize.mapwizecomponents.ui.UIBehaviour
 import io.mapwize.mapwizeformapbox.api.MapwizeObject
 import io.mapwize.mapwizeformapbox.api.Place
 import io.mapwize.mapwizeformapbox.map.MapOptions
 import io.mapwize.mapwizeformapbox.map.MapwizePlugin
 import kotlinx.android.synthetic.main.activity_main.*
 
-class MainActivity : AppCompatActivity(), MapwizeFragment.OnFragmentInteractionListener, UIBehaviour {
-
+class MainActivity : AppCompatActivity(), MapwizeFragment.OnFragmentInteractionListener {
 
     private var mapwizeFragment: MapwizeFragment? = null
     private var mapboxMap: MapboxMap? = null
@@ -42,7 +41,6 @@ class MainActivity : AppCompatActivity(), MapwizeFragment.OnFragmentInteractionL
                 //.compassHidden(true)
                 .build()
         mapwizeFragment = MapwizeFragment.newInstance(opts, uiSettings)
-        this.mapwizeFragment?.uiBehaviour = this
         val fm = supportFragmentManager
         val ft = fm.beginTransaction()
         ft.add(fragmentContainer.id, mapwizeFragment!!)
@@ -67,19 +65,19 @@ class MainActivity : AppCompatActivity(), MapwizeFragment.OnFragmentInteractionL
     }
 
     override fun onMenuButtonClick() {
-        // TODO Do something with menu click
+
     }
 
-    override fun onInformationButtonClick(place: Place?) {
-        // TODO Do something with information button
+    override fun onInformationButtonClick(mapwizeObject: MapwizeObject?) {
+
     }
 
-    /**
-     * UIBehaviour
-     * MapwizeFragment have a default UIBehaviour. You don't have to implement it if you do not need a custom behaviour.
-     * This implementation is here for demo purpose.
-     */
+    override fun onFollowUserButtonClickWithoutLocation() {
+        Log.i("Debug", "onFollowUserButtonClickWithoutLocation")
+    }
+
     override fun shouldDisplayInformationButton(mapwizeObject: MapwizeObject?): Boolean {
+        Log.i("Debug", "shouldDisplayInformationButton")
         when (mapwizeObject) {
             is Place -> return true
         }
@@ -87,6 +85,7 @@ class MainActivity : AppCompatActivity(), MapwizeFragment.OnFragmentInteractionL
     }
 
     override fun shouldDisplayFloorController(floors: MutableList<Double>?): Boolean {
+        Log.i("Debug", "shouldDisplayFloorController")
         if (floors == null || floors.size <= 1) {
             return false
         }